refactor: move contexts and providers to separate files
