+++ app/articles/history/page.tsx
@@ -1,9 +1,10 @@
 "use client";
 import Loading from "@/app/loading";
 import PageBreadcrumb from "@/components/navigation/pagebreadcrumb";
+import { EcosystemContext } from "@/components/providers/ecosystem-provider";
+import { StorageContext } from "@/components/providers/storage-provider";
 import { Typography } from "@/components/ui/typography";
 import { VersionInfo } from "@/lib/articles/storage";
-import { EcosystemContext, StorageContext } from "@/lib/contexts";
 import { formatTime } from "@/lib/time";
 import Link from "next/link";
 import { notFound, useSearchParams } from "next/navigation";
