+++ components/navigation/optimized-search.tsx
@@ -17,8 +17,9 @@ import { ScrollArea } from "@/components/ui/scroll-area";
 import Anchor from "./anchor";
 import { debounce, optimizedSearch } from "@/lib/search";
 import { cn } from "@/lib/utils";
-import { EcosystemContext, StorageContext } from "@/lib/contexts";
 import { buttonVariants } from "../ui/button";
+import { EcosystemContext } from "@/components/providers/ecosystem-provider";
+import { StorageContext } from "@/components/providers/storage-provider";
 
 export default function OptimizedSearch() {
   const [searchedInput, setSearchedInput] = useState("");
